Renamed enum interface
